Refactored top header into a separate component This will (eventually) avoid all those device specific header offsets
