+++ src/sass/components/_navigator.scss
@@ -9,27 +9,6 @@ $headroom-wrapper-webapp-voter-guide: 104px;
 $headroom-webapp: 104px;
 $header-height-cordova-ballot-ios: -1vh;
 
-// 11/4/18, the following 18 lines might be obsolete and not used
-//$header-height-cordova-network-ios--notch: 4vh;
-//$header-height-cordova-network-ios:       -4vh;
-//$header-height-cordova-android:           -5vh;
-//$headroom-xs-max:           9vh;
-//$headroom-xr:               9vh;
-//$headroom-x:                9vh;
-//$headroom--678-plus:         4vh;
-//$headroom--678:              4vh;
-//$headroom-default:          4vh;
-//$headroom-android-sm:       2vh;
-//$headroom-android-md:       2vh;
-//$headroom-android-lg:       2vh;
-//$headroom-android-xl:       2vh;
-//$headroom-cordova-default:  -5vh;
-//$header-height-cordova-network-ios--ipad:  2vh;
-//$header-height-cordova-network-ios--notch: 5vh;
-//$header-height-cordova-network-ios:       -1vh;
-//$header-height-cordova-android:           -5px;
-
-
 $iphone-x-vertical-spacing: 18px;
 $nav-item-secondary-nav-background-hover-color: darken($page-header-secondary-nav-background-color, 5%);
 $nav-item-secondary-nav-background-active-color: darken($page-header-secondary-nav-background-color, 10%);
